// lib/api/generated/models/userStatusResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ServiceDesk API
 * OpenAPI spec version: 1.0
 */
import type { UserStatusResponseStatus } from './userStatusResponseStatus';

export interface UserStatusResponse {
  availableForAssignment?: boolean;
  status?: UserStatusResponseStatus;
  updatedAt?: string;
}